Move the unit tests out of the `frame` package
The tests in the frame package weren't actually testing anything that
was in that package.
