+++ amqp-impl/src/test/java/io/streamnative/pulsar/handlers/amqp/test/AmqpClientMethodProcessor.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.streamnative.pulsar.handlers.amqp.test.frame;
+package io.streamnative.pulsar.handlers.amqp.test;
 
 import lombok.extern.log4j.Log4j2;
 import org.apache.qpid.server.protocol.ProtocolVersion;
