+++ amqp-impl/src/test/java/io/streamnative/pulsar/handlers/amqp/test/AmqpProtocolTestBase.java
@@ -27,10 +27,6 @@ import io.streamnative.pulsar.handlers.amqp.AmqpServiceConfiguration;
 import io.streamnative.pulsar.handlers.amqp.AmqpTopicManager;
 import io.streamnative.pulsar.handlers.amqp.ExchangeContainer;
 import io.streamnative.pulsar.handlers.amqp.impl.PersistentExchange;
-import io.streamnative.pulsar.handlers.amqp.test.frame.AmqpClientChannel;
-import io.streamnative.pulsar.handlers.amqp.test.frame.AmqpClientMethodProcessor;
-import io.streamnative.pulsar.handlers.amqp.test.frame.ToClientByteBufferSender;
-import io.streamnative.pulsar.handlers.amqp.test.frame.ToServerByteBufferSender;
 import io.streamnative.pulsar.handlers.amqp.test.mock.MockDispatcher;
 import io.streamnative.pulsar.handlers.amqp.test.mock.MockManagedLedger;
 import java.net.SocketAddress;
