+++ amqp-impl/src/test/java/io/streamnative/pulsar/handlers/amqp/test/ToClientByteBufferSender.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.streamnative.pulsar.handlers.amqp.test.frame;
+package io.streamnative.pulsar.handlers.amqp.test;
 
 import io.streamnative.pulsar.handlers.amqp.AmqpByteBufferSender;
 import io.streamnative.pulsar.handlers.amqp.AmqpClientDecoder;
